refactors login into Local and External components
